Moves set_alias command in edly app
